refactor: Update database models and bot logic
Improve database schema and bot interaction by:

- Changing user ID field type to BigIntegerField for better
  compatibility.
- Adding a Payment model to track payment transactions.
- Modifying Lottery model to use timezone-aware default
  register_date.
- Removing redundant print statements from views.py and
  adding error handling to the bot.
- Updating migrations to reflect schema changes.
- Adding a `.gitignore` file for better project management.

+++ monogram/models.py
@@ -2,7 +2,7 @@ import pickle
 from django.db import models
 
 class Conversation(models.Model):
-    user_id = models.PositiveIntegerField()  # User ID as a positive integer
+    user_id = models.BigIntegerField()  # User ID as a positive integer
     callback_data = models.CharField(max_length=255, blank=True, null=True)
     status = models.BooleanField(default=True)  # Conversation open status
